Changed encrypted types to Hibernate 6 Converter custom types.
