+++ jasypt-hibernate6/pom.xml
@@ -25,7 +25,7 @@
   <groupId>org.jasypt</groupId>
   <artifactId>jasypt-hibernate6</artifactId>
   <packaging>jar</packaging>
-  <version>1.9.4-SNAPSHOT</version>
+  <version>2.0.0-SNAPSHOT</version>
   <name>JASYPT: Java Simplified Encryption</name>
   <url>http://www.jasypt.org</url>
   
@@ -214,8 +214,6 @@
     </plugins>
     
   </build>
-
-
 
   <dependencies>
