+++ jasypt-hibernate6/src/main/java/org/jasypt/hibernate6/converters/EncryptedByteAsString.java
@@ -1,0 +1,15 @@
+package org.jasypt.hibernate6.converters;
+
+public class EncryptedByteAsString extends AbstractEncryptedAsString {
+
+    @Override
+    protected Object convertToObject(String string) {
+        return Byte.valueOf(string);
+    }
+
+    @Override
+    protected String convertToString(Object object) {
+        return object.toString();
+    }
+
+}
