+++ jasypt-hibernate6/src/main/java/org/jasypt/hibernate6/converters/EncryptedDoubleAsString.java
@@ -1,0 +1,15 @@
+package org.jasypt.hibernate6.converters;
+
+public class EncryptedDoubleAsString extends AbstractEncryptedAsString {
+
+    @Override
+    protected Object convertToObject(String string) {
+        return Double.parseDouble(string);
+    }
+
+    @Override
+    protected String convertToString(Object object) {
+        return object.toString();
+    }
+
+}
