+++ jasypt-hibernate6/src/main/java/org/jasypt/hibernate6/converters/EncryptedString.java
@@ -1,0 +1,14 @@
+package org.jasypt.hibernate6.converters;
+
+public class EncryptedString extends AbstractEncryptedAsString {
+
+    @Override
+    protected Object convertToObject(String string) {
+        return string;
+    }
+
+    @Override
+    protected String convertToString(Object object) {
+        return object.toString();
+    }
+}
