+++ jasypt-hibernate6/src/main/java/org/jasypt/hibernate6/converters/ParameterNaming.java
@@ -17,7 +17,7 @@
  * 
  * =============================================================================
  */
-package org.jasypt.hibernate6.type;
+package org.jasypt.hibernate6.converters;
 
 /**
  * <p>
