+++ jasypt-hibernate6/src/main/java/org/jasypt/hibernate6/encryptor/HibernatePBEEncryptorRegistry.java
@@ -70,7 +70,7 @@ import org.jasypt.encryption.pbe.PBEStringEncryptor;
  * </p>
  * <p>
  * <pre>
- *    &lt;typedef name="encryptedString" class="org.jasypt.hibernate.type.EncryptedStringType">
+ *    &lt;typedef name="encryptedString" class="org.jasypt.hibernate.type.EncryptedString">
  *      &lt;param name="encryptorRegisteredName"><b>myHibernateEncryptor</b>&lt;/param>
  *    &lt;/typedef>
  * </pre>
